Fix warning

+++ Dev/SourcesDemos/DemoGame/Actors/Characters/Character.cpp
@@ -163,7 +163,7 @@ void Character::Update(const DeltaTime& dt)
         m_lifeBar->SetValue(m_currentLife, m_maxLife);
     }
 
-    m_sprite->SetZIndex(m_sprite->GetPosition().y);
+    m_sprite->SetZIndex((int32)m_sprite->GetPosition().y);
 }
 
 sf::Vector2f Character::GetPosition() const
